[ICU-535] fixed some compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/2234

+++ icu4c/source/i18n/sortkey.cpp
@@ -318,7 +318,7 @@ CollationKey::copyValues(int32_t &size) const
     {
         while (output < size)
         {
-            result[output] = (input[0] << 8) | input[1];
+            result[output] = (uint16_t)((input[0] << 8) | input[1]);
             output += 1;
             input += 2;
         }
